made email sending service workable and checked with some email , fixed the issues in clients and projects ..
