+++ elintegrostartapp/grails-app/domain/com/elintegro/website/ClientProject.groovy
@@ -1,4 +1,4 @@
-package elintegroWebsite
+package com.elintegro.website
 
 class ClientProject {
