+++ elintegrostartapp/grails-app/domain/com/elintegro/website/ContactUs.groovy
@@ -1,6 +1,12 @@
-package elintegroWebsite
+package com.elintegro.website
 
 class ContactUs {
+    String name
+    String email
+    String phone
+    String textOfMessage
+    Integer sendNo=0
+    boolean resend = true
 
     static constraints = {
         name(nullable:false)
@@ -9,10 +15,4 @@ class ContactUs {
         textOfMessage(nullable:false)
 
     }
-    String name
-    String email
-    String phone
-    String textOfMessage
-    Integer sendNo=0
-    boolean resend
 }
