+++ elintegrostartapp/grails-app/i18n/messages.properties
@@ -329,4 +329,8 @@ vueElintegroNavigationButtonDataframe.button.contactUs= Contact Us
 vueElintegroNavigationButtonDataframe.button.login= Login
 vueElintegroNavigationButtonDataframe.button.register= Register
 
-textOfMessage = Text Of Message
+Clientname = Client Name 
+Projectname = Project Name 
+LinkToWebsite = Link To Website
+
+vueContactUsPageDataframe.contactUs.textOfMessage = Text Of Message
