+++ elintegrostartapp/grails-app/jobs/com/elintegro/website/contactUsMessageJob.groovy
@@ -1,9 +1,9 @@
-package elintegroWebsite
+package com.elintegro.website
 
 class contactUsMessageJob {
     def contactUsMessageService
     static triggers = {
-        simple repeatInterval: 20000
+        simple repeatInterval: 1*60*60*1000 //send email once in 5 minutes
     }
 
     def execute() {
