+++ elintegrostartapp/src/test/groovy/com/elintegro/website/ElintegroWebsiteControllerSpec.groovy
@@ -1,4 +1,4 @@
-package elintegroWebsite
+package com.elintegro.website
 
 import grails.testing.web.controllers.ControllerUnitTest
 import spock.lang.Specification
